1. Fix bug. 2. Add ProcessTest. 3. Add ApplicationTest.

+++ src/net/Dns.cpp
@@ -226,7 +226,7 @@ namespace Net {
     void Dns::getHostAddresses(IPAddresses &addresses) {
         StringArray ifaces;
         NetInterface::getInterfaceNames(ifaces);
-        for (int i = 0; i < ifaces.count(); ++i) {
+        for (size_t i = 0; i < ifaces.count(); ++i) {
             const String &iface = ifaces[i];
             IPAddress address = NetInterface::getIpAddress(iface);
             if (!address.isEmpty()) {
